ref: Move ChangesRequireFullSyncDialog to own file
