+++ ankihub/gui/operations/ankihub_sync.py
@@ -10,19 +10,7 @@ import aqt.sync
 from anki.collection import OpChangesWithCount
 from anki.hooks import wrap
 from anki.sync import SyncOutput, SyncStatus
-from aqt.qt import (
-    QCheckBox,
-    QDialog,
-    QFrame,
-    QHBoxLayout,
-    QLabel,
-    QPushButton,
-    QSizePolicy,
-    Qt,
-    QTextEdit,
-    QVBoxLayout,
-    qconnect,
-)
+from aqt.qt import qconnect
 from aqt.sync import get_sync_status
 
 from ... import LOGGER
@@ -34,14 +22,10 @@ from ...main.exceptions import ChangesRequireFullSyncError
 from ...main.review_data import send_daily_review_summaries, send_review_data
 from ...main.utils import collection_schema
 from ...settings import config, get_end_cutoff_date_for_sending_review_summaries
+from ..changes_require_full_sync_dialog import ChangesRequireFullSyncDialog
 from ..deck_updater import ah_deck_updater, show_tooltip_about_last_deck_updates_results
 from ..exceptions import FullSyncCancelled
-from ..utils import (
-    CollapsibleSection,
-    logged_into_ankiweb,
-    sync_with_ankiweb,
-    using_qt5,
-)
+from ..utils import logged_into_ankiweb, sync_with_ankiweb
 from .db_check import maybe_check_databases
 from .new_deck_subscriptions import check_and_install_new_deck_subscriptions
 from .utils import future_with_exception, future_with_result, pass_exceptions_to_on_done
@@ -53,119 +37,6 @@ class _SyncState:
 
 
 sync_state = _SyncState()
-
-
-class ChangesRequireFullSyncDialog(QDialog):
-    def __init__(
-        self,
-        changes_require_full_sync_error: ChangesRequireFullSyncError,
-        parent,
-    ):
-        super().__init__(parent)
-        self.setWindowTitle(" ")
-        self.setMinimumWidth(400)
-
-        main_layout = QVBoxLayout(self)
-        main_layout.setContentsMargins(20, 20, 20, 20)
-        main_layout.setSpacing(8)
-
-        # Title inside the dialog
-        title_label = QLabel("<h3>Some changes require a full sync</h3>")
-        title_label.setWordWrap(True)
-        main_layout.addWidget(title_label)
-        main_layout.addSpacing(20)
-
-        # Collapsible Note Type Updates Section, with a maximum expanded height.
-        collapsible = CollapsibleSection("Note type updates", expanded_max_height=160)
-        collapsible.toggle_button.setStyleSheet(
-            collapsible.toggle_button.styleSheet() + "QToolButton { color: gray; }"
-        )
-
-        # Layout for the collapsible content
-        content_layout = QVBoxLayout()
-        content_layout.setContentsMargins(5, 5, 5, 5)
-        content_layout.setSpacing(5)
-
-        self.note_updates_text = QTextEdit()
-        self.note_updates_text.setText(
-            "\n".join(
-                aqt.mw.col.models.get(mid)["name"]
-                for mid in changes_require_full_sync_error.affected_note_type_ids
-            )
-        )
-        self.note_updates_text.setReadOnly(True)
-        self.note_updates_text.setVerticalScrollBarPolicy(
-            Qt.ScrollBarPolicy.ScrollBarAsNeeded
-        )
-        content_layout.addWidget(self.note_updates_text)
-
-        collapsible.setContentLayout(content_layout)
-        main_layout.addWidget(collapsible)
-        main_layout.addSpacing(20)
-
-        # Warning label
-        warning_label = QLabel(
-            "⚠️ <b>Prevent data loss:</b> make sure all your devices are synced with AnkiWeb before proceeding."
-        )
-        warning_label.setWordWrap(True)
-        main_layout.addWidget(warning_label)
-        main_layout.addSpacing(10)
-
-        # Checkbox to enable the full sync button
-        self.synced_checkbox = QCheckBox("I have synced my devices")
-        main_layout.addWidget(self.synced_checkbox)
-        main_layout.addSpacing(20)
-
-        # Separator
-        separator = QFrame()
-        separator.setFrameShape(QFrame.Shape.HLine)
-        separator.setFrameShadow(QFrame.Shadow.Sunken)
-        main_layout.addWidget(separator)
-        main_layout.addSpacing(20)
-
-        # Mobile instructions label
-        mobile_instructions = QLabel(
-            "👉 <b>On mobile</b>, after full sync, select the appropriate option when prompted:"
-            "<ul>"
-            "<li><b>iOS</b>: “Download from AnkiWeb”</li>"
-            "<li><b>Android</b>: “AnkiWeb” or “Keep AnkiWeb collection”<br></li>"
-            "</ul>"
-        )
-        mobile_instructions.setWordWrap(True)
-        main_layout.addWidget(mobile_instructions)
-
-        main_layout.addStretch()
-
-        # Buttons layout
-        button_layout = QHBoxLayout()
-        button_layout.addStretch()
-        skip_button = QPushButton("Skip for now")
-        run_full_sync_button = QPushButton("Run Full Sync")
-        run_full_sync_button.setEnabled(False)
-
-        skip_button.setSizePolicy(QSizePolicy.Policy.Fixed, QSizePolicy.Policy.Fixed)
-        run_full_sync_button.setSizePolicy(
-            QSizePolicy.Policy.Fixed, QSizePolicy.Policy.Fixed
-        )
-
-        button_layout.addWidget(skip_button)
-        button_layout.addWidget(run_full_sync_button)
-        main_layout.addLayout(button_layout)
-
-        # Enable/disable the Run Full Sync button based on checkbox state.
-        qconnect(
-            self.synced_checkbox.stateChanged
-            if using_qt5()
-            else self.synced_checkbox.checkStateChanged,
-            lambda state: run_full_sync_button.setEnabled(
-                state == Qt.CheckState.Checked
-            ),
-        )
-
-        qconnect(skip_button.clicked, self.reject)
-        qconnect(run_full_sync_button.clicked, self.accept)
-
-        self.setLayout(main_layout)
 
 
 @pass_exceptions_to_on_done
